requirements.txt
gym
numpy
torch
pyyaml
torchvision
matplotlib  